requirements.txt
fastapi
uvicorn
pydantic
firebase_admin
pyrebase4
requests